refactor: updated components names and removed old comments
